Extract UIRefreshControlSpy to a separate file
