+++ EssentialFeediOSTests/Helpers/UIRefreshControlSpy.swift
@@ -1,0 +1,20 @@
+//
+// Created by Rodrigo Porto.
+// Copyright © 2024 PortoCode. All Rights Reserved.
+//
+
+import UIKit
+
+class UIRefreshControlSpy: UIRefreshControl {
+    private var _isRefreshing = false
+    
+    override var isRefreshing: Bool { _isRefreshing }
+    
+    override func beginRefreshing() {
+        _isRefreshing = true
+    }
+    
+    override func endRefreshing() {
+        _isRefreshing = false
+    }
+}
